requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
wordcloud
google-generativeai
